requirements.txt
httpx
beautifulsoup4
mcp
pytesseract
pdf2image
pillow
pdfminer.six
